Fixed bug where player wouldn't spawn in correct location between levels
Every level blueprint checks to see whether the player's respawn location variable is at zero. This is how the checkpoints are able to work properly.
However, this would cause the player to spawn at 0 in the next level after completing the previous one.
Thus, I set the end of level door to reset the respawn location once the player has crossed it.
Thereby making the respawn location 0 which is the condition for the level setting it to be where the player start is.